Conflicts:
	include/openspace/engine/openspaceengine.h
	src/engine/openspaceengine.cpp
	src/engine/settingsengine.cpp
	src/engine/syncengine.cpp
	src/interaction/interactionhandler.cpp
	src/rendering/renderengine.cpp
	src/scene/scene.cpp
	src/scene/scenegraph.cpp
	src/scene/scenegraphnode.cpp
	tests/test_scenegraphloader.inl

+++ modules/volume/volumeutils.h
@@ -2,7 +2,7 @@
  *                                                                                       *
  * OpenSpace                                                                             *
  *                                                                                       *
- * Copyright (c) 2014-2016                                                               *
+ * Copyright (c) 2014-2017                                                               *
  *                                                                                       *
  * Permission is hereby granted, free of charge, to any person obtaining a copy of this  *
  * software and associated documentation files (the "Software"), to deal in the Software *
@@ -25,13 +25,13 @@
 #ifndef __OPENSPACE_MODULE_VOLUME___VOLUMEUTILS___H__
 #define __OPENSPACE_MODULE_VOLUME___VOLUMEUTILS___H__
 
-#include <glm/glm.hpp>
+#include <ghoul/glm.h>
 
 namespace openspace {
 namespace volumeutils {
 
-size_t coordsToIndex(const glm::vec3& coords, const glm::ivec3& dimensions);
-glm::vec3 indexToCoords(size_t index, const glm::ivec3& dimensions);
+size_t coordsToIndex(const glm::uvec3& coords, const glm::uvec3& dimensions);
+glm::uvec3 indexToCoords(size_t index, const glm::uvec3& dimensions);
 
 } // namespace volumeutils
